pyproject.toml
[tool.poetry]
name = "survivor-predict"
version = "0.1.0"
description = ""
authors = ["Jared Lunde <jared.lunde@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pychalk = "^2.0"
scipy = "^1.3"
python-dotenv = "^0.10.3"
requests = "^2.22"
diskcache = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
